# restAPI/src/main/swagger/openapi.yaml
openapi: 3.0.0
info:
  description: This is the API for the UI's access to data storage
  version: 1.0.0
  title: AIDA Store REST API
servers:
  - url: "{protocol}://{host}:{port}/{basePath}"
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      host:
        default: localhost
      port:
        default: "8008"
      basePath:
        default: api
...
paths:
...
  /events/types:
    get:
      tags:
        - Event
      summary: Return event types matching provided input
      description: Return event types by graph and filter by specified input
      operationId: eventTypes
      parameters:
        - $ref: "#/components/parameters/graphParam"
        - name: type
          in: query
          description: Part or all of event type
          required: false
          schema:
            type: string
          example: Conflict.Attack
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "400":
          description: Invalid parameters
        "404":
          description: API not found
  /events/arguments:
    get:
      tags:
        - Event
      summary: Return event argument names matching provided input
      description: Return event argument names by graph and filter by specified input
      operationId: eventNames
      parameters:
        - $ref: "#/components/parameters/graphParam"
        - $ref: "#/components/parameters/argumentParam"
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "400":
          description: Invalid parameters
        "404":
          description: API not found
...
  /event/element:
    get:
      tags:
        - Event
      summary: Return a single event element
      description: Return event element by id
      operationId: eventElement
      parameters:
        - name: id
          in: query
          description: ID of event element to retrieve
          required: true
          schema:
            type: string
          example: http://www.lti.cs.cmu.edu/aida/opera/corpora/eval/event-instance-IC001V5AJ-r202010162251-28
        - $ref: "#/components/parameters/optionalGraphParam"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Element"
        "400":
          description: Invalid parameters
        "404":
          description: API not found
...
components:
  schemas:
...
    Member:
      allOf:
        - $ref: "#/components/schemas/JustifiedNode"
        - type: object
          properties:
            arguments:
              type: array
              items:
                $ref: "#/components/schemas/Role"
    Element:
      allOf:
        - $ref: "#/components/schemas/Member"
        - type: object
          properties:
            clusters:
              type: array
              items:
                $ref: "#/components/schemas/JustifiedNode"
    Role:
      type: object
      properties:
        role:
          type: string
        fillers:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
    JustifiedNode:
      type: object
      discriminator:
        propertyName: objectType
      properties:
        id:
          type: string
        handle:
          type: string
        objectType:
          type: string
          enum: ["Entity", "Member", "Element", "EntityMember", "EntityElement", "JustifiedNode"]
        types:
          type: array
          items:
            type: string
        docs:
          type: array
          items:
            $ref: "#/components/schemas/DocObject"
    Entity:
      allOf:
        - $ref: "#/components/schemas/JustifiedNode"
        - type: object
          properties:
            names:
              type: array
              items:
                type: string
            category:
              type: string
            clusterIds:
              type: array
              items:
                type: string
...
    DocObject:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        contentDate:
          type: string
        downloadDate:
          type: string
...
  parameters:
    graphParam:
      name: graph
      in: query
      description: Which graph to query in
      required: true
      schema:
        type: string
      example: https://www.nextcentury.com/TA2/OPERA-20210215/OPERA-20210117
    optionalGraphParam:
      name: graph
      in: query
      description: Which graph to query in
      schema:
        type: string
      example: https://www.nextcentury.com/TA2/OPERA-20210215/OPERA-20210117
...
    argumentParam:
      name: argument
      in: query
      description: Part or all of argument name
      required: false
      schema:
        type: string
      example: Chavez
    limitParam:
      name: dataLimit
      in: query
      description: The limit of the number of data returned.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 500
...